Add script to parse and save flattented ntuple to JSON
This takes all the tree parsing & data collection from
plotCompareNtuples.py, so that the data can be extracted
separately from plotting & comparing it

Note that for space reasons, we concatenate the data
from all events together, so each method string has
one long list of values. That's because we don't care
about individual events later on.

It is saved as JSON because trying to save as a ROOT TTree
in PyROOT is painful (how to do variable size branches?)
Saving in hdf5 occupies more space.